+++ tests/test_pattern.py
@@ -2,7 +2,7 @@ import unittest
 import numpy as np
 import tests.random_circuit as rc
 from graphix.transpiler import Circuit
-from graphix.pattern import Pattern
+from graphix.pattern import Pattern, CommandNode
 
 
 class TestPattern(unittest.TestCase):
@@ -12,7 +12,8 @@ class TestPattern(unittest.TestCase):
         pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
         circuit = rc.generate_gate(nqubits, depth, pairs)
         pattern = circuit.transpile()
-        pattern.standardize()
+        pattern.standardize(method="global")
+        np.testing.assert_equal(pattern.is_standard(), True)
         state = circuit.simulate_statevector()
         state_mbqc = pattern.simulate_pattern()
         np.testing.assert_almost_equal(np.abs(np.dot(state_mbqc.flatten().conjugate(), state.flatten())), 1)
@@ -23,6 +24,7 @@ class TestPattern(unittest.TestCase):
         pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
         circuit = rc.generate_gate(nqubits, depth, pairs)
         pattern = circuit.transpile()
+        pattern.standardize(method="global")
         pattern.minimize_space()
         state = circuit.simulate_statevector()
         state_mbqc = pattern.simulate_pattern()
@@ -34,7 +36,8 @@ class TestPattern(unittest.TestCase):
         pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
         circuit = rc.generate_gate(nqubits, depth, pairs)
         pattern = circuit.transpile()
-        pattern.shift_signals()
+        pattern.standardize(method="global")
+        pattern.shift_signals(method="global")
         pattern.perform_pauli_measurements()
         pattern.minimize_space()
         state = circuit.simulate_statevector()
@@ -48,6 +51,7 @@ class TestPattern(unittest.TestCase):
             pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
             circuit = rc.generate_gate(nqubits, depth, pairs)
             pattern = circuit.transpile()
+            pattern.standardize(method="global")
             pattern.minimize_space()
             np.testing.assert_equal(pattern.max_space(), nqubits + 1)
 
@@ -57,6 +61,7 @@ class TestPattern(unittest.TestCase):
         pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
         circuit = rc.generate_gate(nqubits, depth, pairs)
         pattern = circuit.transpile()
+        pattern.standardize(method="global")
         pattern.parallelize_pattern()
         state = circuit.simulate_statevector()
         state_mbqc = pattern.simulate_pattern()
@@ -68,7 +73,9 @@ class TestPattern(unittest.TestCase):
         pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
         circuit = rc.generate_gate(nqubits, depth, pairs)
         pattern = circuit.transpile()
-        pattern.shift_signals()
+        pattern.standardize(method="global")
+        pattern.shift_signals(method="global")
+        np.testing.assert_equal(pattern.is_standard(), True)
         state = circuit.simulate_statevector()
         state_mbqc = pattern.simulate_pattern()
         np.testing.assert_almost_equal(np.abs(np.dot(state_mbqc.flatten().conjugate(), state.flatten())), 1)
@@ -79,8 +86,8 @@ class TestPattern(unittest.TestCase):
         pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
         circuit = rc.generate_gate(nqubits, depth, pairs)
         pattern = circuit.transpile()
-        pattern.standardize()
-        pattern.shift_signals()
+        pattern.standardize(method="global")
+        pattern.shift_signals(method="global")
         pattern.perform_pauli_measurements()
         pattern.minimize_space()
         state = circuit.simulate_statevector()
@@ -93,8 +100,8 @@ class TestPattern(unittest.TestCase):
         pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
         circuit = rc.generate_gate(nqubits, depth, pairs, use_rzz=True)
         pattern = circuit.transpile(opt=True)
-        pattern.standardize()
-        pattern.shift_signals()
+        pattern.standardize(method="global")
+        pattern.shift_signals(method="global")
         pattern.perform_pauli_measurements()
         pattern.minimize_space()
         state = circuit.simulate_statevector()
@@ -107,8 +114,8 @@ class TestPattern(unittest.TestCase):
         pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
         circuit = rc.generate_gate(nqubits, depth, pairs, use_rzz=True)
         pattern = circuit.standardize_and_transpile(opt=True)
-        pattern.standardize()
-        pattern.shift_signals()
+        pattern.standardize(method="global")
+        pattern.shift_signals(method="global")
         pattern.perform_pauli_measurements()
         pattern.minimize_space()
         state = circuit.simulate_statevector()
@@ -145,8 +152,8 @@ class TestPattern(unittest.TestCase):
         swap(circuit, 0, 2)
 
         pattern = circuit.transpile()
-        pattern.standardize()
-        pattern.shift_signals()
+        pattern.standardize(method="global")
+        pattern.shift_signals(method="global")
         pattern.perform_pauli_measurements()
 
         isolated_nodes = pattern.get_isolated_nodes()
@@ -191,6 +198,200 @@ def swap(circuit, a, b):
     circuit.cnot(a, b)
     circuit.cnot(b, a)
     circuit.cnot(a, b)
+
+
+class TestLocalPattern(unittest.TestCase):
+    def test_assert_equal_edge(self):
+        test_case = [
+            [(0, 1), (0, 1), True],
+            [(1, 0), (0, 1), True],
+            [(0, 1), (2, 0), False],
+            [(0, 1), (2, 3), False],
+            [(1, 3), (4, 1), False],
+        ]
+        for test in test_case:
+            np.testing.assert_equal(assert_equal_edge(test[0], test[1]), test[2])
+
+    def test_no_gate(self):
+        n = 3
+        circuit = Circuit(n)
+        pattern = circuit.transpile()
+        localpattern = pattern.get_local_pattern()
+        for node in localpattern.nodes.values():
+            np.testing.assert_equal(node.seq, [])
+
+    def test_get_graph(self):
+        nqubits = 5
+        depth = 4
+        pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
+        circuit = rc.generate_gate(nqubits, depth, pairs)
+        pattern = circuit.transpile()
+        nodes_ref, edges_ref = pattern.get_graph()
+
+        localpattern = pattern.get_local_pattern()
+        nodes, edges = localpattern.get_graph()
+
+        # nodes check
+        nodes_check1 = True
+        nodes_check2 = True
+        for node in nodes:
+            if node not in nodes_ref:
+                nodes_check1 = False
+        for node in nodes_ref:
+            if node not in nodes:
+                nodes_check2 = False
+        np.testing.assert_equal(nodes_check1, True)
+        np.testing.assert_equal(nodes_check2, True)
+
+        # edges check
+        edges_check1 = True
+        edges_check2 = True
+        for edge in edges:
+            edge_match = False
+            for edge_ref in edges_ref:
+                edge_match |= assert_equal_edge(edge, edge_ref)
+            if not edge_match:
+                edges_check1 = False
+        for edge in edges_ref:
+            edge_match = False
+            for edge_ref in edges:
+                edge_match |= assert_equal_edge(edge, edge_ref)
+            if not edge_match:
+                edges_check2 = False
+        np.testing.assert_equal(edges_check1, True)
+        np.testing.assert_equal(edges_check2, True)
+
+    def test_standardize(self):
+        nqubits = 5
+        depth = 4
+        pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
+        circuit = rc.generate_gate(nqubits, depth, pairs)
+        pattern = circuit.transpile()
+        localpattern = pattern.get_local_pattern()
+        localpattern.standardize()
+        pattern = localpattern.get_pattern()
+        np.testing.assert_equal(pattern.is_standard(), True)
+        pattern.minimize_space()
+        state_p = pattern.simulate_pattern()
+        state_ref = circuit.simulate_statevector()
+        np.testing.assert_almost_equal(np.abs(np.dot(state_p.flatten().conjugate(), state_ref.flatten())), 1)
+
+    def test_shift_signals(self):
+        nqubits = 5
+        depth = 4
+        pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
+        circuit = rc.generate_gate(nqubits, depth, pairs)
+        pattern = circuit.transpile()
+        localpattern = pattern.get_local_pattern()
+        localpattern.standardize()
+        localpattern.shift_signals()
+        pattern = localpattern.get_pattern()
+        np.testing.assert_equal(pattern.is_standard(), True)
+        pattern.minimize_space()
+        state_p = pattern.simulate_pattern()
+        state_ref = circuit.simulate_statevector()
+        np.testing.assert_almost_equal(np.abs(np.dot(state_p.flatten().conjugate(), state_ref.flatten())), 1)
+
+    def test_standardize_and_shift_signals(self):
+        nqubits = 5
+        depth = 4
+        pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
+        circuit = rc.generate_gate(nqubits, depth, pairs)
+        pattern = circuit.transpile()
+        pattern.standardize_and_shift_signals()
+        np.testing.assert_equal(pattern.is_standard(), True)
+        pattern.minimize_space()
+        state_p = pattern.simulate_pattern()
+        state_ref = circuit.simulate_statevector()
+        np.testing.assert_almost_equal(np.abs(np.dot(state_p.flatten().conjugate(), state_ref.flatten())), 1)
+
+    def test_mixed_pattern_operations(self):
+        processes = [
+            [["standardize", "global"], ["standardize", "local"]],
+            [["standardize", "local"], ["signal", "global"], ["signal", "local"]],
+            [
+                ["standardize", "local"],
+                ["signal", "global"],
+                ["standardize", "global"],
+                ["signal", "local"],
+            ],
+        ]
+        nqubits = 3
+        depth = 2
+        pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
+        circuit = rc.generate_gate(nqubits, depth, pairs)
+        state_ref = circuit.simulate_statevector()
+        for process in processes:
+            pattern = circuit.transpile()
+            for operation in process:
+                if operation[0] == "standardize":
+                    pattern.standardize(method=operation[1])
+                elif operation[0] == "signal":
+                    pattern.shift_signals(method=operation[1])
+            np.testing.assert_equal(pattern.is_standard(), True)
+            pattern.minimize_space()
+            state_p = pattern.simulate_pattern()
+            np.testing.assert_almost_equal(np.abs(np.dot(state_p.flatten().conjugate(), state_ref.flatten())), 1)
+
+    def test_opt_transpile_standardize(self):
+        nqubits = 5
+        depth = 4
+        pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
+        circuit = rc.generate_gate(nqubits, depth, pairs)
+        pattern = circuit.transpile(opt=True)
+        pattern.standardize(method="local")
+        np.testing.assert_equal(pattern.is_standard(), True)
+        pattern.minimize_space()
+        state_p = pattern.simulate_pattern()
+        state_ref = circuit.simulate_statevector()
+        np.testing.assert_almost_equal(np.abs(np.dot(state_p.flatten().conjugate(), state_ref.flatten())), 1)
+
+    def test_opt_transpile_shift_signals(self):
+        nqubits = 5
+        depth = 4
+        pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
+        circuit = rc.generate_gate(nqubits, depth, pairs)
+        pattern = circuit.transpile(opt=True)
+        pattern.standardize(method="local")
+        pattern.shift_signals(method="local")
+        np.testing.assert_equal(pattern.is_standard(), True)
+        pattern.minimize_space()
+        state_p = pattern.simulate_pattern()
+        state_ref = circuit.simulate_statevector()
+        np.testing.assert_almost_equal(np.abs(np.dot(state_p.flatten().conjugate(), state_ref.flatten())), 1)
+
+    def test_node_is_standardized(self):
+        ref_sequence = [
+            [[1, 2, 3, -1], True],
+            [[1, 2, 3, -2, -3, -2, -4], True],
+            [[1, -4, 2, -3, -1, 3], False],
+            [[1, 2, 3, -1, -4, 2], False],
+        ]
+        for [seq, ref] in ref_sequence:
+            node = CommandNode(0, seq, [], [], False, [], [])
+            result = node.is_standard()
+            np.testing.assert_equal(result, ref)
+
+    def test_localpattern_is_standard(self):
+        nqubits = 5
+        depth = 4
+        pairs = [(i, np.mod(i + 1, nqubits)) for i in range(nqubits)]
+        circuit = rc.generate_gate(nqubits, depth, pairs)
+        localpattern = circuit.transpile().get_local_pattern()
+        result1 = localpattern.is_standard()
+        localpattern.standardize()
+        result2 = localpattern.is_standard()
+        np.testing.assert_equal(result1, False)
+        np.testing.assert_equal(result2, True)
+
+
+def assert_equal_edge(edge, ref):
+    if (edge[0] == ref[0]) and (edge[1] == ref[1]):
+        return True
+    elif (edge[0] == ref[1]) and (edge[1] == ref[0]):
+        return True
+    else:
+        return False
 
 
 if __name__ == "__main__":