+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Fast pattern standardization and signal shfiting with `pattern.LocalPattern` class (#42), performance report at #43
+- Defaulted local pattern method for `graphix.Pattern.standardize()` and `graphix.Pattern.shift_signals()`. Note the resulting pattern is equivalent to the output of original method.
+
+### Changed
+
+
+## [0.2.1] - 2023-04-25
+
 ### Changed
 
 - Move import path of `generate_from_pattern` from `graphix.gflow` to `grahpix.generator` (#40)